CIRTbuildenv: Split out cyclictest property file creation
The function which prepares and writes the cyclictest environment property
file should be reused by other tests.

As a preparatory step move the creation of the boottest property file into
the calling function.

While at it replace .each() function, which is not necessarily
serializable, with serializable C-style for loop.

No functional change.

Signed-off-by: Benedikt Spranger <b.spranger@linutronix.de>
Signed-off-by: Anna-Maria Gleixner <anna-maria@linutronix.de>