checkoutKernel: Introduce LINUX_KERNEL_MIRROR Jenkins node environment variable
To speed up the kernel checkout a kernel source mirror is recommended.
Introduce LINUX_KERNEL_MIRROR environment variable to point to a
kernel source mirror.

LINUX_KERNEL_MIRROR can be set up in the node specific environment setup,
i.e. it is a Jenkins node environment variable, to be capable to handle
different locations of the kernel source mirror on each node.

Signed-off-by: Benedikt Spranger <b.spranger@linutronix.de>
Signed-off-by: Anna-Maria Gleixner <anna-maria@linutronix.de>