boottest2xml: use optionparser and introduce --skip
boottests are either sucessfully exectuded, failed or are skipped if
the DUT is not available.

Instead of marking the test as skipped in junit, the testcase will be
currently not recorded.

This extends boottest2xml with a --skip parameter to record a boottest
as skipped.

Instead of making things more complicated with hand-written option
parsing convert code and callers to use OptionParse.

Signed-off-by: Manuel Traut <manut@linutronix.de>
Reviewed-by: Benedikt Spranger <b.spranger@linutronix.de>