compiletestRunner: Use external gitdescribe script and improve error handling
To be able to test short parts of the compiletest stand-alone, the commands
required for creation of GIT_TAG and GIT_NAME are splitted into a short bash
script.

While at it, improve error handling and complain, when GIT_NAME stays empty
because no tags are available and complain also when an unknown error occured.

Signed-off-by: Anna-Maria Gleixner <anna-maria@linutronix.de>
Reviewed-by: Benedikt Spranger <b.spranger@linutronix.de>