gittags.properties: Create this file only once after kernel checkout
gittags.properties file was created on every compiletest run and added to the
test results of the compiletest. This file is required only once and there is no
difference, if this command is executed after applying the quilt queue.

Therefore the gittags.properties file creation is moved into checkoutKernel and
is stored in a stash. All places, where the information of the gittags file is
required could use now the stash.

Signed-off-by: Anna-Maria Gleixner <anna-maria@linutronix.de>
Reviewed-by: Benedikt Spranger <b.spranger@linutronix.de>