CIRTscheduler: Do not execute stages on errors in checkout
When there was a known error like VarNotSetExecption in stage 'checkout kernel
source', stages 'compile' and 'collect and inform' should not be
executed. Therefore introduce a variable which is checked at the begin of the
stages.

The reason why a variable is used, which is able to abort only 'script'
sections, is that the 'when' statement for the stage is able to read environment
or job parameters only. It is not able to read variables. environment is valid
only for a single stage, so it could not be used for our problem. parameters are
immutable so this wouldn't be a solution either.

Signed-off-by: Anna-Maria Gleixner <anna-maria@linutronix.de>
Reviewed-by: Benedikt Spranger <b.spranger@linutronix.de>