resources/.../gitdescribe.sh: Introduce configurable '--no-tags' git option
When fetching the history the tags are fetched as well. For regression testing
when forcing a NoGitTagsException there needs to be the possibility to prevent
fetching tags.

Therefore introduce a script argument which specifies whether git option
'--no-tags' should be used. The callside of the script is adapted as well and
calls the script with false right now. So no functional change at the moment.

Signed-off-by: Anna-Maria Gleixner <anna-maria@linutronix.de>
Reviewed-by: Benedikt Spranger <b.spranger@linutronix.de>