checkoutKernel: Enable REGRESSION_TEST_NOGITTAGS handling
To force the exception NoGitTagsException, the git calls are all executed with
'--no-tags' option. In gitCheckout() and gitdescribe.sh, the optional
configuration is already in place. To enable this behavior, test-description
needs to contain REGRESSION_TEST_NOGITTAGS=true and it is required, that it is
parsed before executing git commands.

Signed-off-by: Anna-Maria Gleixner <anna-maria@linutronix.de>
Reviewed-by: Benedikt Spranger <b.spranger@linutronix.de>