feedDatabase: Set PostgreSQL database schema
The default PostgreSQL database schema "public" allows authentificated
users to add, modify and remove database tables by default. This
special schema was added not to break databases created before the
schema support landed into PostgreSQL. While reworking database
permissions in CI-RT there are two options:

1) remove permissions granted by default
2) add a explicit schema

The later options was used not to break the default behaviour of
PostgreSQL.

ATTENTION: This patch needs a database update and will not work without the
changes in the database.

Signed-off-by: Benedikt Spranger <b.spranger@linutronix.de>
Signed-off-by: Anna-Maria Gleixner <anna-maria@linutronix.de>