generictest: Notify on failure
If a generictest fails the user can directly be informed. The generictest
is executed under pyjutest to generate a JUnit XML file. This file is used
to carry the test result and is analyzed later on. Unfortionally
redirecting stdout and stderr does not work as expected. Therefore generate
a log file from the JUnit XML file in case of an error. This logfile is
used to inform the user in a human readable way and do not fluster kernel
developers with XML content.

Signed-off-by: Benedikt Spranger <b.spranger@linutronix.de>
Signed-off-by: Anna-Maria Gleixner <anna-maria@linutronix.de>