Remove ensureDebPkg
ensureDebPkg tries to purge and reinstall a package if it is not installed
or installed in an other version. This make system updates to a
force majeure. ensureDebPkg is only used by pyjutest. Failures on a non
existing package are reliable and better to debug, than some abitrary
apt --purge command.
Drop ensureDebPkg entirely.

Signed-off-by: Benedikt Spranger <b.spranger@linutronix.de>
Reviewed-by: Bastian Germann <bage@linutronix.de>