boottestRunner: Attach raw boot log on failure
The boot log contain kernel messages with a priority of emerg, alert,
crit or err. That is quite enough to track errors between different
test runs. But it is far too less to debug a failure. Attach the
unfiltered raw boot log to the failure notification.
While at it compress the log to minimize the email footprint.

Signed-off-by: Benedikt Spranger <b.spranger@linutronix.de>
Reviewed-by: Bastian Germann <bage@linutronix.de>